+++ database/migrations/2013_10_11_091142_create_areas_table.php
@@ -14,11 +14,11 @@ class CreateAreasTable extends Migration
     public function up()
     {
         Schema::create('areas', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('ar_name');
             $table->string('ar_description')->nullable();
             $table->string('ar_keywords', 20)->nullable();
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
             $table->timestamps();
             $table->softDeletes();
