+++ database/migrations/2021_08_31_111435_create_subject_status_table.php
@@ -16,17 +16,17 @@ class CreateSubjectStatusTable extends Migration
         
 
         Schema::create('subject_status', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name', 255)->nullable();
             $table->string('description', 255)->nullable();
             //$table->string('type', 255)->nullable();
 
-            $table->integer('type_matter_id')->unsigned();
+            $table->bigInteger('type_matter_id')->unsigned();
             $table->foreign('type_matter_id')->references('id')->on('type_subjects');
            
             
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->timestamps();
