+++ database/migrations/2021_08_31_112714_create_subjects_table.php
@@ -14,7 +14,7 @@ class CreateSubjectsTable extends Migration
     public function up()
     {
         Schema::create('subjects', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('mat_name', 255)->nullable();
             $table->string('cod_matter_migration', 255)->nullable();
@@ -24,16 +24,16 @@ class CreateSubjectsTable extends Migration
             $table->string('mat_description', 255)->nullable();
             $table->enum('mat_payment_type', ['horas', 'creditos', 'Ninguno']);
 
-            $table->integer('type_matter_id')->unsigned();
+            $table->bigInteger('type_matter_id')->unsigned();
             $table->foreign('type_matter_id')->references('id')->on('type_subjects');
 
-            $table->integer('education_level_id')->unsigned();
+            $table->bigInteger('education_level_id')->unsigned();
             $table->foreign('education_level_id')->references('id')->on('education_levels');
 
-            $table->integer('area_id')->unsigned();
+            $table->bigInteger('area_id')->unsigned();
             $table->foreign('area_id')->references('id')->on('areas');
 
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->timestamps();
