+++ database/migrations/2021_08_31_113504_create_pensums_table.php
@@ -14,13 +14,13 @@ class CreatePensumsTable extends Migration
     public function up()
     {
         Schema::create('pensums', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('pen_name', 255)->nullable();
             $table->string('pen_description', 255)->nullable();
             $table->string('pen_acronym', 3)->nullable();
-            $table->integer('anio');
+            $table->bigInteger('anio');
 
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->timestamps();
