+++ database/migrations/2021_08_31_113648_create_type_periods_table.php
@@ -14,18 +14,18 @@ class CreateTypePeriodsTable extends Migration
     public function up()
     {
         Schema::create('type_periods', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('tp_name', 255)->nullable();
             $table->string('tp_description', 255)->nullable();
 
-            $table->integer('tp_min_matter_enrollment')->nullable();
-            $table->integer('tp_max_matter_enrollment')->nullable();
-            $table->integer('tp_num_fees')->nullable();
-            $table->integer('tp_fees_enrollment')->nullable();
+            $table->bigInteger('tp_min_matter_enrollment')->nullable();
+            $table->bigInteger('tp_max_matter_enrollment')->nullable();
+            $table->bigInteger('tp_num_fees')->nullable();
+            $table->bigInteger('tp_fees_enrollment')->nullable();
             $table->boolean('tp_pay_enrollment');
 
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->timestamps();
