+++ database/migrations/2021_08_31_130504_create_table_collaborator.php
@@ -15,28 +15,28 @@ class CreateTableCollaborator extends Migration
     {
         Schema::create('collaborators', function (Blueprint $table) {
             // Add the Auto-Increment column
-            //$table->increments("coll_contract_num");
+            //$table->bigIncrements("coll_contract_num");
 
             // Remove the primary key
             //$table->dropPrimary("collaborators_coll_contract_num_primary");
 
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->string('coll_email')->unique();
             $table->string('coll_type',1)->nullable();
             $table->string('coll_activity')->nullable();
-            $table->integer('coll_journey_hours')->nullable();
+            $table->bigInteger('coll_journey_hours')->nullable();
             $table->string('coll_journey_description')->nullable();
 
             $table->boolean('coll_dependency')->nullable();
             $table->boolean('coll_advice')->nullable(); //campo booleano para saber si es consejero
 
-            $table->integer('position_company_id')->unsigned();
+            $table->bigInteger('position_company_id')->unsigned();
             $table->foreign('position_company_id')->references('id')->on('positions');
 
-            $table->integer('education_level_principal_id')->unsigned();
+            $table->bigInteger('education_level_principal_id')->unsigned();
             $table->foreign('education_level_principal_id')->references('id')->on('education_levels');
 
             $table->timestamp('coll_entering_date')->nullable();
@@ -48,7 +48,7 @@ class CreateTableCollaborator extends Migration
             $table->timestamp('coll_nomination_leaving_date')->nullable();
             $table->timestamp('coll_disabled_reason')->nullable();
 
-            $table->integer('status_id')->unsigned();
+            $table->bigInteger('status_id')->unsigned();
             $table->foreign('status_id')->references('id')->on('status');
 
             $table->timestamps();
